Login,Register,Major Changes
1. Implemented token-Authentication with budget app adding login and registration form
2. Error Handling added for forms, register form shows error from Django
3. changed from localstorage to sessionStorage so automatically deletes token when window is closed.
4. Added a logout button to remove token from sessionStorage.
5. Implement separate layouts for login/register and main content.
6. Implemented AuthGuard to prevent unauthorized access to pages(without token)
7. Implemeted refresh token, so session cannot be expired.
8. Added proper input validation to all the form pages.
9. Fixed issue with loading transaction list after deleting one.
10. Created a landing page to navigate after login.
11. Added a pop up before deleting a content.

+++ src/app/listing/budget-list.ts
@@ -52,6 +52,8 @@ export class BudgetList implements OnInit{
   }
 
   deleteBudget(budget: any) {
+    const confirmed = confirm(`Do you really want to delete this budget?`);
+    if (!confirmed) return;
     this.api.deleteBudget(budget.id).subscribe({
       next: () => this.loadBudgets(),
       error: (err) => console.error('Delete failed', err)
